Int was changed to float in line 6 and 7

+++ field_area.py
@@ -2,9 +2,9 @@
 length = input("Enter a length in feet: ")
 width = input("Enter a width in feet: ")
 
-# converting length and width to int
-length = int(length)
-width = int(width)
+# converting length and width to float (to allow for decimals in input)
+length = float(length)
+width = float(width)
 
 # calculate area 
 area = length * width / 43560
